# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/accounts/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a account",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiberswagger.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiberswagger.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiberswagger.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiberswagger.HTTPError"
                        }
                    }
                }
            }
        },
        "/create": {
            "post": {
                "description": "add by customer data json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Add a customer",
                "parameters": [
                    {
                        "description": "Add customer",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CustomerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.CustomerData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Account": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "controller.CustomerData": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birthdate": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idcard": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "tel": {
                    "type": "string"
                }
            }
        },
        "controller.CustomerInput": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string",
                    "format": "date",
                    "example": "2000-01-02"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mark.woodsome@gmail.com"
                },
                "firstname": {
                    "type": "string",
                    "format": "first name",
                    "example": "mark"
                },
                "idcard": {
                    "type": "string",
                    "format": "id card",
                    "example": "11129292929292"
                },
                "lastname": {
                    "type": "string",
                    "format": "last name",
                    "example": "woodsome"
                },
                "tel": {
                    "type": "string",
                    "format": "mobile no.",
                    "example": "0877766988"
                }
            }
        },
        "controller.HTTPError": {
            "type": "object"
        },
        "fiberswagger.Account": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "fiberswagger.HTTPError": {
            "type": "object"
        },
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}